+++ tardigrade/roots/aws/baseline/variables.tf
@@ -113,3 +113,40 @@ variable "config_bucket_versioning" {
   description = "Whether versioning is enabled on parent config bucket"
   default     = false
 }
+
+##### SECURITY GROUP VARIABLES #####
+variable "default_vpc_sg_ingress_rules" {
+  description = "A schema list of ingress rules for the default vpc's default security group, see https://www.terraform.io/docs/providers/aws/r/security_group.html#ingress"
+  type        = list
+  default     = []
+}
+
+variable "default_vpc_sg_egress_rules" {
+  description = "A schema list of egress rules for the default vpc's default security group, see https://www.terraform.io/docs/providers/aws/r/security_group.html#egress"
+  type        = list
+  default     = []
+}
+
+variable "default_vpc_revoke_rules_on_delete" {
+  description = "Determines whether to forcibly remove rules when destroying the default vpc's default security group"
+  type        = string
+  default     = false
+}
+
+variable "vpc_module_sg_ingress_rules" {
+  description = "A schema list of ingress rules for the vpc module's default security group, see https://www.terraform.io/docs/providers/aws/r/security_group.html#ingress"
+  type        = list
+  default     = []
+}
+
+variable "vpc_module_sg_egress_rules" {
+  description = "A schema list of egress rules for the vpc module's default security group, see https://www.terraform.io/docs/providers/aws/r/security_group.html#egress"
+  type        = list
+  default     = []
+}
+
+variable "vpc_module_revoke_rules_on_delete" {
+  description = "Determines whether to forcibly remove rules when destroying the vpc module's default security group"
+  type        = string
+  default     = false
+}