+++ CHANGELOG.md
@@ -4,9 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### 0.0.2
+
+**Commit Delta**: [Change from 0.0.1 release](https://github.com/plus3it/tardigrade/compare/0.0.1...0.0.2)
+
+**Released**: 2020.01.02
+
+**Summary**:
+
+*   Update cloudtrail version which integrates kms
+*   Remove rules from default vpc's default security group
+*   Remove 0.0.0.0/0 rule from the created vpc's default security group
+
 ### 0.0.1
 
-**Commit Delta**: 
+**Commit Delta**: [Change from 0.0.0 release](https://github.com/plus3it/tardigrade/compare/0.0.0...0.0.1)
 
 **Released**: 2019.10.28